refactor(project-pattern): context hook and types
