+++ src/types/contexts/themeContextTypes.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react'
+
+export interface ThemeProviderProps {
+  children: ReactNode
+}
